Produce upgrade zip

+++ dev/scripts/release.py
@@ -2,6 +2,8 @@ import subprocess
 from subprocess import PIPE
 import itertools
 import shutil
+from pathlib import Path
+import sys
 
 
 EXCLUDE_DIRS = [
@@ -38,7 +40,7 @@ EXCLUDE_FILES = [
 TARGET_DIR = 'release'
 
 
-def create_archives(archive_name: str):
+def create_archives(source_dir: str, archive_name: str):
     zip_command = [
         'zip',
         '-r',  # Recursive
@@ -71,22 +73,49 @@ def create_archives(archive_name: str):
 
 
 if __name__ == '__main__':
+    if not Path('.git').exists():
+        print('.git does not exists')
+        sys.exit(1)
+
     print('Deleting vendor files')
     shutil.rmtree('core/assets/vendor', ignore_errors=True)
     shutil.rmtree('node_modules', ignore_errors=True)
     shutil.rmtree('vendor', ignore_errors=True)
 
     # Create base archive
-    create_archives('base')
+    create_archives('.', 'base')
 
     # Run npm and composer (production dependencies only)
     subprocess.check_call(['npm', 'ci', '-q', '--cache', '.node_cache'],
                           stdout=PIPE)
     subprocess.check_call(['composer', 'install', '--no-dev', '--no-interaction'],
                           stdout=PIPE)
-    create_archives('deps-dist')
+    create_archives('.', 'deps-dist')
 
     # Run composer again, to install development dependencies
     subprocess.check_call(['composer', 'install', '--no-interaction'],
                           stdout=PIPE)
-    create_archives('deps-dev')
+    create_archives('.', 'deps-dev')
+
+    # Create archive with files changed since last update
+
+    upgrade_temp = Path('release', 'upgrade_temp')
+
+    # Find previous tag
+    previous_tag_command = ['git', 'describe', '--abbrev=0', '--tags']
+    previous_tag = subprocess.check_output(previous_tag_command, shell=False)[:-1].decode()
+
+    # Find all files changed between previous tag and HEAD (current state)
+    changed_command = ['git', 'diff', previous_tag, 'HEAD', '--name-only', '--diff-filter=d']
+    changed_files = subprocess.check_output(changed_command, shell=False)[:-1].decode()
+
+    for changed_file in changed_files.split('\n'):
+        changed_file_target = Path(upgrade_temp, changed_file)
+        changed_file_target.parent.mkdir(parents=True, exist_ok=True)
+        shutil.copy2(changed_file, changed_file_target)
+
+    # Vendor files are always included
+    for vendor_dir in ['vendor', 'core/assets/vendor']:
+        shutil.copytree(vendor_dir, Path(upgrade_temp, vendor_dir))
+
+    create_archives(upgrade_temp.as_posix(), 'upgrade-from-' + previous_tag)
